Re-organized files and folders that they follow MVP architecture pattern
